Move shred iterator to its' own file
